refactor(cli): split check guidance module
